+++ protocols/relay/src/v2/client/handler.rs
@@ -31,12 +31,12 @@ use instant::Instant;
 use libp2p_core::either::EitherError;
 use libp2p_core::multiaddr::Protocol;
 use libp2p_core::{upgrade, ConnectedPoint, Multiaddr, PeerId};
-use libp2p_swarm::protocols_handler::{
-    DummyProtocolsHandler, InboundUpgradeSend, OutboundUpgradeSend, SendWrapper,
+use libp2p_swarm::handler::{
+    DummyConnectionHandler, InboundUpgradeSend, OutboundUpgradeSend, SendWrapper,
 };
 use libp2p_swarm::{
-    IntoProtocolsHandler, KeepAlive, NegotiatedSubstream, ProtocolsHandler, ProtocolsHandlerEvent,
-    ProtocolsHandlerUpgrErr, SubstreamProtocol,
+    ConnectionHandler, ConnectionHandlerEvent, ConnectionHandlerUpgrErr, IntoConnectionHandler,
+    KeepAlive, NegotiatedSubstream, SubstreamProtocol,
 };
 use log::debug;
 use std::collections::VecDeque;
@@ -79,12 +79,12 @@ pub enum Event {
     ReservationReqFailed {
         /// Indicates whether the request replaces an existing reservation.
         renewal: bool,
-        error: ProtocolsHandlerUpgrErr<outbound_hop::ReservationFailedReason>,
+        error: ConnectionHandlerUpgrErr<outbound_hop::ReservationFailedReason>,
     },
     /// An outbound circuit has been established.
     OutboundCircuitEstablished { limit: Option<protocol::Limit> },
     OutboundCircuitReqFailed {
-        error: ProtocolsHandlerUpgrErr<outbound_hop::CircuitFailedReason>,
+        error: ConnectionHandlerUpgrErr<outbound_hop::CircuitFailedReason>,
     },
     /// An inbound circuit has been established.
     InboundCircuitEstablished {
@@ -93,7 +93,7 @@ pub enum Event {
     },
     /// An inbound circuit request has failed.
     InboundCircuitReqFailed {
-        error: ProtocolsHandlerUpgrErr<void::Void>,
+        error: ConnectionHandlerUpgrErr<void::Void>,
     },
     /// An inbound circuit request has been denied.
     InboundCircuitReqDenied { src_peer_id: PeerId },
@@ -119,8 +119,8 @@ impl Prototype {
     }
 }
 
-impl IntoProtocolsHandler for Prototype {
-    type Handler = Either<Handler, DummyProtocolsHandler>;
+impl IntoConnectionHandler for Prototype {
+    type Handler = Either<Handler, DummyConnectionHandler>;
 
     fn into_handler(self, remote_peer_id: &PeerId, endpoint: &ConnectedPoint) -> Self::Handler {
         if endpoint.is_relayed() {
@@ -133,7 +133,7 @@ impl IntoProtocolsHandler for Prototype {
             }
 
             // Deny all substreams on relayed connection.
-            Either::Right(DummyProtocolsHandler::default())
+            Either::Right(DummyConnectionHandler::default())
         } else {
             let mut handler = Handler {
                 remote_peer_id: *remote_peer_id,
@@ -156,7 +156,7 @@ impl IntoProtocolsHandler for Prototype {
         }
     }
 
-    fn inbound_protocol(&self) -> <Self::Handler as ProtocolsHandler>::InboundProtocol {
+    fn inbound_protocol(&self) -> <Self::Handler as ConnectionHandler>::InboundProtocol {
         upgrade::EitherUpgrade::A(SendWrapper(inbound_stop::Upgrade {}))
     }
 }
@@ -167,7 +167,7 @@ pub struct Handler {
     remote_addr: Multiaddr,
     /// A pending fatal error that results in the connection being closed.
     pending_error: Option<
-        ProtocolsHandlerUpgrErr<
+        ConnectionHandlerUpgrErr<
             EitherError<inbound_stop::FatalUpgradeError, outbound_hop::FatalUpgradeError>,
         >,
     >,
@@ -176,11 +176,11 @@ pub struct Handler {
 
     /// Queue of events to return when polled.
     queued_events: VecDeque<
-        ProtocolsHandlerEvent<
-            <Self as ProtocolsHandler>::OutboundProtocol,
-            <Self as ProtocolsHandler>::OutboundOpenInfo,
-            <Self as ProtocolsHandler>::OutEvent,
-            <Self as ProtocolsHandler>::Error,
+        ConnectionHandlerEvent<
+            <Self as ConnectionHandler>::OutboundProtocol,
+            <Self as ConnectionHandler>::OutboundOpenInfo,
+            <Self as ConnectionHandler>::OutEvent,
+            <Self as ConnectionHandler>::Error,
         >,
     >,
 
@@ -201,14 +201,14 @@ pub struct Handler {
     /// Futures that try to send errors to the transport.
     ///
     /// We may drop errors if this handler ends up in a terminal state (by returning
-    /// [`ProtocolsHandlerEvent::Close`]).
+    /// [`ConnectionHandlerEvent::Close`]).
     send_error_futs: FuturesUnordered<BoxFuture<'static, ()>>,
 }
 
-impl ProtocolsHandler for Handler {
+impl ConnectionHandler for Handler {
     type InEvent = In;
     type OutEvent = Event;
-    type Error = ProtocolsHandlerUpgrErr<
+    type Error = ConnectionHandlerUpgrErr<
         EitherError<inbound_stop::FatalUpgradeError, outbound_hop::FatalUpgradeError>,
     >;
     type InboundProtocol = inbound_stop::Upgrade;
@@ -242,7 +242,7 @@ impl ProtocolsHandler for Handler {
                     relay_addr: self.remote_addr.clone(),
                 });
 
-                self.queued_events.push_back(ProtocolsHandlerEvent::Custom(
+                self.queued_events.push_back(ConnectionHandlerEvent::Custom(
                     Event::InboundCircuitEstablished {
                         src_peer_id: self.remote_peer_id,
                         limit,
@@ -285,7 +285,7 @@ impl ProtocolsHandler for Handler {
                 );
 
                 self.queued_events
-                    .push_back(ProtocolsHandlerEvent::Custom(event));
+                    .push_back(ConnectionHandlerEvent::Custom(event));
             }
 
             // Outbound circuit
@@ -305,7 +305,7 @@ impl ProtocolsHandler for Handler {
                 ))) {
                     Ok(()) => {
                         self.alive_lend_out_substreams.push(rx);
-                        self.queued_events.push_back(ProtocolsHandlerEvent::Custom(
+                        self.queued_events.push_back(ConnectionHandlerEvent::Custom(
                             Event::OutboundCircuitEstablished { limit },
                         ));
                     }
@@ -325,7 +325,7 @@ impl ProtocolsHandler for Handler {
         match event {
             In::Reserve { to_listener } => {
                 self.queued_events
-                    .push_back(ProtocolsHandlerEvent::OutboundSubstreamRequest {
+                    .push_back(ConnectionHandlerEvent::OutboundSubstreamRequest {
                         protocol: SubstreamProtocol::new(
                             outbound_hop::Upgrade::Reserve,
                             OutboundOpenInfo::Reserve { to_listener },
@@ -337,7 +337,7 @@ impl ProtocolsHandler for Handler {
                 dst_peer_id,
             } => {
                 self.queued_events
-                    .push_back(ProtocolsHandlerEvent::OutboundSubstreamRequest {
+                    .push_back(ConnectionHandlerEvent::OutboundSubstreamRequest {
                         protocol: SubstreamProtocol::new(
                             outbound_hop::Upgrade::Connect { dst_peer_id },
                             OutboundOpenInfo::Connect { send_back },
@@ -350,35 +350,35 @@ impl ProtocolsHandler for Handler {
     fn inject_listen_upgrade_error(
         &mut self,
         _: Self::InboundOpenInfo,
-        error: ProtocolsHandlerUpgrErr<<Self::InboundProtocol as InboundUpgradeSend>::Error>,
+        error: ConnectionHandlerUpgrErr<<Self::InboundProtocol as InboundUpgradeSend>::Error>,
     ) {
         let non_fatal_error = match error {
-            ProtocolsHandlerUpgrErr::Timeout => ProtocolsHandlerUpgrErr::Timeout,
-            ProtocolsHandlerUpgrErr::Timer => ProtocolsHandlerUpgrErr::Timer,
-            ProtocolsHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Select(
+            ConnectionHandlerUpgrErr::Timeout => ConnectionHandlerUpgrErr::Timeout,
+            ConnectionHandlerUpgrErr::Timer => ConnectionHandlerUpgrErr::Timer,
+            ConnectionHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Select(
                 upgrade::NegotiationError::Failed,
-            )) => ProtocolsHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Select(
+            )) => ConnectionHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Select(
                 upgrade::NegotiationError::Failed,
             )),
-            ProtocolsHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Select(
+            ConnectionHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Select(
                 upgrade::NegotiationError::ProtocolError(e),
             )) => {
-                self.pending_error = Some(ProtocolsHandlerUpgrErr::Upgrade(
+                self.pending_error = Some(ConnectionHandlerUpgrErr::Upgrade(
                     upgrade::UpgradeError::Select(upgrade::NegotiationError::ProtocolError(e)),
                 ));
                 return;
             }
-            ProtocolsHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Apply(
+            ConnectionHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Apply(
                 inbound_stop::UpgradeError::Fatal(error),
             )) => {
-                self.pending_error = Some(ProtocolsHandlerUpgrErr::Upgrade(
+                self.pending_error = Some(ConnectionHandlerUpgrErr::Upgrade(
                     upgrade::UpgradeError::Apply(EitherError::A(error)),
                 ));
                 return;
             }
         };
 
-        self.queued_events.push_back(ProtocolsHandlerEvent::Custom(
+        self.queued_events.push_back(ConnectionHandlerEvent::Custom(
             Event::InboundCircuitReqFailed {
                 error: non_fatal_error,
             },
@@ -388,38 +388,38 @@ impl ProtocolsHandler for Handler {
     fn inject_dial_upgrade_error(
         &mut self,
         open_info: Self::OutboundOpenInfo,
-        error: ProtocolsHandlerUpgrErr<<Self::OutboundProtocol as OutboundUpgradeSend>::Error>,
+        error: ConnectionHandlerUpgrErr<<Self::OutboundProtocol as OutboundUpgradeSend>::Error>,
     ) {
         match open_info {
             OutboundOpenInfo::Reserve { mut to_listener } => {
                 let non_fatal_error = match error {
-                    ProtocolsHandlerUpgrErr::Timeout => ProtocolsHandlerUpgrErr::Timeout,
-                    ProtocolsHandlerUpgrErr::Timer => ProtocolsHandlerUpgrErr::Timer,
-                    ProtocolsHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Select(
+                    ConnectionHandlerUpgrErr::Timeout => ConnectionHandlerUpgrErr::Timeout,
+                    ConnectionHandlerUpgrErr::Timer => ConnectionHandlerUpgrErr::Timer,
+                    ConnectionHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Select(
                         upgrade::NegotiationError::Failed,
-                    )) => ProtocolsHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Select(
+                    )) => ConnectionHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Select(
                         upgrade::NegotiationError::Failed,
                     )),
-                    ProtocolsHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Select(
+                    ConnectionHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Select(
                         upgrade::NegotiationError::ProtocolError(e),
                     )) => {
-                        self.pending_error = Some(ProtocolsHandlerUpgrErr::Upgrade(
+                        self.pending_error = Some(ConnectionHandlerUpgrErr::Upgrade(
                             upgrade::UpgradeError::Select(
                                 upgrade::NegotiationError::ProtocolError(e),
                             ),
                         ));
                         return;
                     }
-                    ProtocolsHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Apply(error)) => {
+                    ConnectionHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Apply(error)) => {
                         match error {
                             outbound_hop::UpgradeError::Fatal(error) => {
-                                self.pending_error = Some(ProtocolsHandlerUpgrErr::Upgrade(
+                                self.pending_error = Some(ConnectionHandlerUpgrErr::Upgrade(
                                     upgrade::UpgradeError::Apply(EitherError::B(error)),
                                 ));
                                 return;
                             }
                             outbound_hop::UpgradeError::ReservationFailed(error) => {
-                                ProtocolsHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Apply(
+                                ConnectionHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Apply(
                                     error,
                                 ))
                             }
@@ -447,7 +447,7 @@ impl ProtocolsHandler for Handler {
                 }
 
                 let renewal = self.reservation.failed();
-                self.queued_events.push_back(ProtocolsHandlerEvent::Custom(
+                self.queued_events.push_back(ConnectionHandlerEvent::Custom(
                     Event::ReservationReqFailed {
                         renewal,
                         error: non_fatal_error,
@@ -456,33 +456,33 @@ impl ProtocolsHandler for Handler {
             }
             OutboundOpenInfo::Connect { send_back } => {
                 let non_fatal_error = match error {
-                    ProtocolsHandlerUpgrErr::Timeout => ProtocolsHandlerUpgrErr::Timeout,
-                    ProtocolsHandlerUpgrErr::Timer => ProtocolsHandlerUpgrErr::Timer,
-                    ProtocolsHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Select(
+                    ConnectionHandlerUpgrErr::Timeout => ConnectionHandlerUpgrErr::Timeout,
+                    ConnectionHandlerUpgrErr::Timer => ConnectionHandlerUpgrErr::Timer,
+                    ConnectionHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Select(
                         upgrade::NegotiationError::Failed,
-                    )) => ProtocolsHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Select(
+                    )) => ConnectionHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Select(
                         upgrade::NegotiationError::Failed,
                     )),
-                    ProtocolsHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Select(
+                    ConnectionHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Select(
                         upgrade::NegotiationError::ProtocolError(e),
                     )) => {
-                        self.pending_error = Some(ProtocolsHandlerUpgrErr::Upgrade(
+                        self.pending_error = Some(ConnectionHandlerUpgrErr::Upgrade(
                             upgrade::UpgradeError::Select(
                                 upgrade::NegotiationError::ProtocolError(e),
                             ),
                         ));
                         return;
                     }
-                    ProtocolsHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Apply(error)) => {
+                    ConnectionHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Apply(error)) => {
                         match error {
                             outbound_hop::UpgradeError::Fatal(error) => {
-                                self.pending_error = Some(ProtocolsHandlerUpgrErr::Upgrade(
+                                self.pending_error = Some(ConnectionHandlerUpgrErr::Upgrade(
                                     upgrade::UpgradeError::Apply(EitherError::B(error)),
                                 ));
                                 return;
                             }
                             outbound_hop::UpgradeError::CircuitFailed(error) => {
-                                ProtocolsHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Apply(
+                                ConnectionHandlerUpgrErr::Upgrade(upgrade::UpgradeError::Apply(
                                     error,
                                 ))
                             }
@@ -497,7 +497,7 @@ impl ProtocolsHandler for Handler {
 
                 let _ = send_back.send(Err(()));
 
-                self.queued_events.push_back(ProtocolsHandlerEvent::Custom(
+                self.queued_events.push_back(ConnectionHandlerEvent::Custom(
                     Event::OutboundCircuitReqFailed {
                         error: non_fatal_error,
                     },
@@ -514,7 +514,7 @@ impl ProtocolsHandler for Handler {
         &mut self,
         cx: &mut Context<'_>,
     ) -> Poll<
-        ProtocolsHandlerEvent<
+        ConnectionHandlerEvent<
             Self::OutboundProtocol,
             Self::OutboundOpenInfo,
             Self::OutEvent,
@@ -524,7 +524,7 @@ impl ProtocolsHandler for Handler {
         // Check for a pending (fatal) error.
         if let Some(err) = self.pending_error.take() {
             // The handler will not be polled again by the `Swarm`.
-            return Poll::Ready(ProtocolsHandlerEvent::Close(err));
+            return Poll::Ready(ConnectionHandlerEvent::Close(err));
         }
 
         // Return queued events.
@@ -533,7 +533,7 @@ impl ProtocolsHandler for Handler {
         }
 
         if let Poll::Ready(Some(protocol)) = self.reservation.poll(cx) {
-            return Poll::Ready(ProtocolsHandlerEvent::OutboundSubstreamRequest { protocol });
+            return Poll::Ready(ConnectionHandlerEvent::OutboundSubstreamRequest { protocol });
         }
 
         // Deny incoming circuit requests.
@@ -541,12 +541,12 @@ impl ProtocolsHandler for Handler {
         {
             match result {
                 Ok(()) => {
-                    return Poll::Ready(ProtocolsHandlerEvent::Custom(
+                    return Poll::Ready(ConnectionHandlerEvent::Custom(
                         Event::InboundCircuitReqDenied { src_peer_id },
                     ))
                 }
                 Err(error) => {
-                    return Poll::Ready(ProtocolsHandlerEvent::Custom(
+                    return Poll::Ready(ConnectionHandlerEvent::Custom(
                         Event::InboundCircuitReqDenyFailed { src_peer_id, error },
                     ))
                 }