+++ protocols/autonat/CHANGELOG.md
@@ -1,4 +1,10 @@
-# 0.2.0 [unreleased]
+# 0.3.0 [unreleased]
+
+- Update to `libp2p-swarm` `v0.35.0`.
+
+- Update to `libp2p-request-response` `v0.17.0`.
+
+# 0.2.0 [2022-02-22]
 
 - Update to `libp2p-core` `v0.32.0`.